backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
transformers==4.37.0
torch==2.1.2
accelerate==0.26.0
bitsandbytes==0.42.0
pillow==10.2.0
python-multipart==0.0.6
pytest==8.0.0
httpx==0.26.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
cryptography==41.0.7
redis==5.0.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-cache2[redis]==0.2.1
slowapi==0.1.9
pyinstrument==4.6.2
locust==2.24.0
